RegiowizardBot/requirements.txt
streamlit
openai
langchain
langchain_community
pypdf
sentence_transformers
python-dotenv
chromadb
torch
transformers
langdetect
